cloud functions initialized
Also there is an error whenever I click "Proceed to checkout" or remove from basket at payment page.
Note: This might be due to our implementation of navigate.replace('/orders', {replace: true}) as it does not match with the function used in the vid history.replace('orders').
Maybe navigate.replace('/orders', {state}) could be the fix, or something else entierly.
It could also be something related to the axios implementation as the error does state itself to be an AxiosError -
AxiosError: Request failed with status code 404